Greyed out countries and changes to colour space
Greys out countries that do not have any data.

Also, the color space was previously always based on the range of our values. This was problematic when no country was selected since the base value is zero in that case. For example, if the range was 0.5-0.7 and no country was selected everything would appear to be completely red without any shading since the colour space would be values (-0.1)-0.1 which does not contain the value space.