More visually appealing hover stroke
Changes to how the hover stroke is drawn. Now it is drawn after all the countries have been drawn and it's stroke size is manipulatable through Javascript. This is done by having invisible copies of the countries with the stroke that are only shown when they are hovered over. This comes at a performance cost when react updates are run but I think it may be worth the cost.